requirements.txt
forexconnect
flask
python-dotenv==0.21.1
numpy==1.14.5
pandas==0.23.4
python-dateutil==2.7.3
pytz==2018.5
six==1.11.0